+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/Settings.java
@@ -43,6 +43,8 @@ public class Settings extends BaseSettings {
     public static final BooleanSetting EDGE_SEEK = new BooleanSetting("edge_seek", FALSE);
     public static final BooleanSetting FIT_VIDEO_TO_SCREEN =
             new BooleanSetting("fit_video_to_screen", FALSE);
+    public static final BooleanSetting UNCAP_REFRESH_RATE =
+            new BooleanSetting("uncap_refresh_rate", FALSE);
     public static final BooleanSetting HIDE_FOLLOWER_NOTIFICATIONS =
             new BooleanSetting("hide_follower_notifications", FALSE);
     public static final BooleanSetting HIDE_MESSAGE_STREAKS =