+++ extensions/tiktok/src/main/java/app/morphe/extension/tiktok/settings/SettingsStatus.java
@@ -63,6 +63,7 @@ public class SettingsStatus {
     }
     public static boolean inboxFilterEnabled = false;
     public static boolean videoFitEnabled = false;
+    public static boolean refreshRateEnabled = false;
     public static boolean notificationControlsEnabled = false;
     public static boolean hideSuggestedAccountsEnabled = false;
     public static boolean hideInboxStoriesEnabled = false;
@@ -185,6 +186,10 @@ public class SettingsStatus {
 
     public static void enableExpandActivityList() {
         expandActivityListEnabled = true;
+    }
+
+    public static void enableRefreshRate() {
+        refreshRateEnabled = true;
     }
 
     public static void enableVideoFit() {